Fix to use MATE logo for dark theme
When using the MATE Faenze Dark Icon theme, the MATE logo next to the Applications menu is from matefaenza/places/start-here.png. This fix should now replace it with the MATE logo found in the matefaezadark/places/start-here.png, which is more suitable for a dark theme. The fix involves adding the places/22 and places/24 directories into the index.theme file.